Separate source map generation into a mixin
